Use new implementations in client code, fix code so that tests pass
The MerginProject object was moved verbatim to the new file without any changes to it.
