+++ mergin/common.py
@@ -1,0 +1,25 @@
+
+
+CHUNK_SIZE = 100 * 1024 * 1024
+
+# there is an upper limit for chunk size on server, ideally should be requested from there once implemented
+UPLOAD_CHUNK_SIZE = 10 * 1024 * 1024
+
+
+class ClientError(Exception):
+    pass
+
+
+try:
+    import dateutil.parser
+    from dateutil.tz import tzlocal
+except ImportError:
+    # this is to import all dependencies shipped with package (e.g. to use in qgis-plugin)
+    deps_dir = os.path.join(this_dir, 'deps')
+    if os.path.exists(deps_dir):
+        import sys
+        for f in os.listdir(os.path.join(deps_dir)):
+            sys.path.append(os.path.join(deps_dir, f))
+
+        import dateutil.parser
+        from dateutil.tz import tzlocal
